refactor(tauri): decompose monolithic commands.rs into modular submodules
