+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,20 @@
+mod apps;
+mod audio_files;
+mod debug;
+mod devices;
+mod helpers;
+mod network;
+mod pipeline;
+mod plugins;
+mod updater;
+mod virtual_devices;
+
+pub use apps::*;
+pub use audio_files::*;
+pub use debug::*;
+pub use devices::*;
+pub use network::*;
+pub use pipeline::*;
+pub use plugins::*;
+pub use updater::*;
+pub use virtual_devices::*;
